Extracted auto retry tests
